fix(markdown): Restore API compatibility
